Flexible, uniform, centralized configuration.
Now:

    Contacts.configure(
      :google => {:consumer_key => ..., :consumer_secret, ...},
      :yahoo => {:consumer_key => ..., :consumer_secret, ...},
      ...
    )

This means we only need to read the configuration once, and we can
configure by any means necessary. e.g., in a Rails app, we may want
different credentials per environment for Yahoo! so we can redirect to
a different domain.

The top two levels of keys may be symbols or strings, for easy YAML
configuration.

Defining instance methods in service classes to retrieve configuration
settings is now also a one-liner. Example:

    configurable_attribute :consumer_key